feat(gateway): CSDL - make `fields` directive args parseable (#4489)
As an artifact of composition, CSDL should generally be easily consumed by
tooling. The existing format for a FieldSet looks like so: @key(fields: "id").
Currently, whenever we make use of that @key in existing (related)
implementations, we parse the fields argument, but not before adding a
surrounding { ... } set of curly braces to make the string valid for parsing by
the graphql library.

This is fine in JS, but it's an additional step. In Rust (where we're consuming
this format for the first time), modifying strings like this rather than making
use of the original string slice is not nearly as trivial due to lifetimes and
ownership. By making this string parseable as-is, we eliminate a step for anyone
using the CSDL format in tooling and eliminate a number of ergonomic issues for
our Rust use case.